requirements.txt
Pillow>=10.0.0
pyinstaller>=6.0.0
pytest>=8.0.0